Deleted long from my variable and replaed it wth int

+++ 0x03-more_functions_nested_loops/101-print_number.c
@@ -9,7 +9,7 @@
 
 void print_number(int n)
 {
-	signed long m = n;
+	signed int m = n;
 
 	if (m < 0)
 	{
